+++ src/Products/PloneMeeting/setuphandlers.py
@@ -380,6 +380,9 @@ def postInstall(context):
         name='Products.CMFPlone.interfaces.syndication.ISiteSyndicationSettings.search_rss_enabled',
         value=False)
 
+    # configure portal_repository
+    _configurePortalRepository()
+
     # configure dexterity localrolesfield
     _configureDexterityLocalRolesField()
 
@@ -586,6 +589,18 @@ def _configure_zamqp(site):
         site.portal_setup.runAllImportStepsFromProfile(
             'imio.zamqp.pm:default',
             dependency_strategy=DEPENDENCY_STRATEGY_REAPPLY)
+
+
+def _configurePortalRepository():
+    """Make sure default Plone content type are not versionable."""
+    plone_versioned_types = [u'ATDocument', u'ATNewsItem', u'Document', u'Event', u'Link', u'News Item']
+    pr = api.portal.get_tool('portal_repository')
+    pr.setVersionableContentTypes([pt_id for pt_id in pr.getVersionableContentTypes()
+                                   if pt_id not in plone_versioned_types])
+    for p_type in plone_versioned_types:
+        pr._version_policy_mapping.pop(p_type, None)
+        # _version_policy_mapping is a dict, make change persistent
+        pr._version_policy_mapping = pr._version_policy_mapping
 
 
 def _configureDexterityLocalRolesField():